+++ class_v2/ssl_dnsV2/pdns_script.sh
@@ -1,16 +1,44 @@
 #!/bin/bash
 PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin
 export PATH
-install_tmp='/tmp/bt_install.pl'
 public_file=/www/server/panel/install/public.sh
 
 if [ ! -f $public_file ];then
-	wget -O $public_file http://download.bt.cn/install/public.sh -T 5;
+	wget -O $public_file https://node.aapanel.com/install/public.sh -T 30;
 fi
 
 . $public_file
 download_Url=$NODE_URL
-pluginPath=/www/server/panel/plugin/dns_manager
+
+
+check_and_disable_resolved()
+{
+  if systemctl is-active --quiet systemd-resolved; then
+    echo "Found active systemd-resolved. Stopping and disabling it..."
+    systemctl stop systemd-resolved
+    systemctl disable systemd-resolved
+    echo "systemd-resolved has been stopped and disabled."
+  elif systemctl list-units --type=service --all | grep -q 'systemd-resolved.service'; then
+    echo "Found inactive systemd-resolved. Disabling it..."
+    systemctl disable systemd-resolved
+    echo "systemd-resolved has been disabled."
+  else
+    echo "systemd-resolved service not found, no action needed."
+  fi
+
+  # 删软链
+  if [ -L /etc/resolv.conf ]; then
+        echo "Removing symlink /etc/resolv.conf"
+        rm -f /etc/resolv.conf
+        echo "Set namerser 8.8.8.8 to /etc/resolv.conf"
+        echo "nameserver 8.8.8.8" > /etc/resolv.conf
+  fi
+  # 确保上游dns
+  if ! grep -q "nameserver 8.8.8.8" /etc/resolv.conf 2>/dev/null; then
+      echo "Adding 8.8.8.8 to /etc/resolv.conf"
+      echo "nameserver 8.8.8.8" >> /etc/resolv.conf
+  fi
+}
 
 Install_Powerdns_Redhat()
 {
@@ -18,7 +46,7 @@ Install_Powerdns_Redhat()
   groupadd named
   useradd -g named -s /sbin/nologin named
   mv /etc/pdns/pdns.conf /etc/pdns/pdns.conf_bt
-  wget -O /etc/pdns/pdns.conf $download_Url/install/plugin/dns_manager/pdns.conf -T 5
+  wget -O /etc/pdns/pdns.conf $download_Url/install/plugin/dns_manager/pdns.conf -T 30
   chmod 644 /etc/pdns/pdns.conf
   mkdir -p /var/named/chroot/etc
   mkdir -p /var/named/chroot/var/named
@@ -36,8 +64,7 @@ Install_Powerdns_Redhat()
   if [ "$bind_conf" == "" ];then
     sed -i 's/file\s*\"/file \"\/var\/named\/chroot\/var\/named\//g' /var/named/chroot/etc/named.rfc1912.zones
   fi
-  systemctl stop systemd-resolved
-  systemctl disable systemd-resolved
+  check_and_disable_resolved
 	systemctl stop named-chroot
 	systemctl disable named-chroot
   systemctl enable pdns
@@ -49,26 +76,20 @@ Install_Powerdns_Redhat()
 
 Install_Powerdns_Ubuntu()
 {
-  ubuntu=$(cat /etc/issue)
   curl https://repo.powerdns.com/FD380FBB-pub.asc | sudo apt-key add -
-  apt-get update
+  apt-get update -y
   apt-get install pdns-server -y
-  if [[ "$ubuntu" =~ "Ubuntu" ]];then
-    systemctl disable systemd-resolved
-    systemctl stop systemd-resolved
-    rm /etc/resolv.conf
-    echo "nameserver 8.8.8.8" > /etc/resolv.conf
-  fi
   groupadd named
   useradd -g named -s /sbin/nologin named
   mv /etc/powerdns/pdns.conf /etc/powerdns/pdns.conf_bt
-  wget -O /etc/powerdns/pdns.conf $download_Url/install/plugin/dns_manager/pdns.conf -T 5
+  wget -O /etc/powerdns/pdns.conf $download_Url/install/plugin/dns_manager/pdns.conf -T 30
   chmod 644 /etc/powerdns/pdns.conf
   mkdir -p /var/named/chroot/etc
   mkdir -p /var/named/chroot/var/named
   if [ ! -f "/var/named/chroot/etc/named.rfc1912.zones" ];then
     touch /var/named/chroot/etc/named.rfc1912.zones
   fi
+  check_and_disable_resolved
   systemctl enable pdns
   systemctl restart pdns
 
@@ -77,74 +98,67 @@ Install_Powerdns_Ubuntu()
 
 Install_Powerdns()
 {
-  if [ -f '/etc/redhat-release' ];then
+  if [ -f '/usr/bin/yum' ];then
     Install_Powerdns_Redhat
   else
     Install_Powerdns_Ubuntu
   fi
+
+  echo -n "pdns" > /www/server/panel/class_v2/ssl_dnsV2/aadns.pl
 }
 
 Install_DnsManager()
 {
-  rm -f $pluginPath/dns_manager_main.py
-	mkdir -p $pluginPath
+	echo "Installing..."
 	Install_Powerdns
-  echo "pdns" > /www/server/panel/class_v2/ssl_dnsV2/aadns.pl
-	/usr/bin/pip install dnspython
-	echo 'Installing...' > $install_tmp
+	/usr/bin/btpip install dnspython
 	grep "English" /www/server/panel/config/config.json
-	if [ "$?" -ne 0 ];then
-        wget -O $pluginPath/dns_manager_main.py $download_Url/install/plugin/dns_manager/dns_manager_main.py -T 5
-        wget -O $pluginPath/index.html $download_Url/install/plugin/dns_manager/index.html -T 5
-        wget -O $pluginPath/info.json $download_Url/install/plugin/dns_manager/info.json -T 5
-        wget -O $pluginPath/icon.png $download_Url/install/plugin/dns_manager/icon.png -T 5
-    else
-        wget -O $pluginPath/dns_manager_main.py $download_Url/install/plugin/dns_manager_en/dns_manager_main.py -T 5
-        wget -O $pluginPath/index.html $download_Url/install/plugin/dns_manager_en/index.html -T 5
-        wget -O $pluginPath/info.json $download_Url/install/plugin/dns_manager_en/info.json -T 5
-        wget -O $pluginPath/icon.png $download_Url/install/plugin/dns_manager_en/icon.png -T 5
-	fi
-    \cp -a -r /www/server/panel/plugin/dns_manager/icon.png /www/server/panel/BTPanel/static/img/soft_ico/ico-dns_manager.png
-
-	echo 'Success' > $install_tmp
-  echo "Installed Success!"
+	sleep 3
+	echo "aaPanelDns Service Success!"
 }
 
 update_DnsManager()
 {
   if [ ! -f "/usr/sbin/pdns_server" ];then
+	echo "Installing ..."
   Install_Powerdns
   fi
-	echo 'Installing ...' > $install_tmp
 	grep "English" /www/server/panel/config/config.json
-	if [ "$?" -ne 0 ];then
-        wget -O $pluginPath/dns_manager_main.py $download_Url/install/plugin/dns_manager/dns_manager_main.py -T 5
-        wget -O $pluginPath/index.html $download_Url/install/plugin/dns_manager/index.html -T 5
-        wget -O $pluginPath/info.json $download_Url/install/plugin/dns_manager/info.json -T 5
-        wget -O $pluginPath/icon.png $download_Url/install/plugin/dns_manager/icon.png -T 5
-    else
-        wget -O $pluginPath/dns_manager_main.py $download_Url/install/plugin/dns_manager_en/dns_manager_main.py -T 5
-        wget -O $pluginPath/index.html $download_Url/install/plugin/dns_manager_en/index.html -T 5
-        wget -O $pluginPath/info.json $download_Url/install/plugin/dns_manager_en/info.json -T 5
-        wget -O $pluginPath/icon.png $download_Url/install/plugin/dns_manager_en/icon.png -T 5
-	fi
-    \cp -a -r /www/server/panel/plugin/dns_manager/icon.png /www/server/panel/static/img/soft_ico/ico-dns_manager.png
-
-	echo 'The installation is complete' > $install_tmp
+	echo "The installation is complete"
 }
 
 Uninstall_DnsManager()
 {
-	rm -rf $pluginPath
-	rm -f /var/named/chroot/etc/named.rfc1912.zones_bak
-	mv /var/named/chroot/etc/named.rfc1912.zones /var/named/chroot/etc/named.rfc1912.zones_bak
-  mv /var/named/chroot/var/named /var/named/chroot/var/named_bak
-  rm -rf /var/named/chroot/var/named
+  echo "Uninstalling..."
+  clean=${1}
+  if [ "$clean" == 1 ];then
+    echo "Cleaning up aaPanelDns config data..."
+    rm -f /var/named/chroot/etc/named.rfc1912.zones_bak
+    mv /var/named/chroot/etc/named.rfc1912.zones /var/named/chroot/etc/named.rfc1912.zones_bak
+    echo "Remove config: named.rfc1912.zones"
+    cp /var/named/chroot/var/named/*_aadef /var/named/chroot/var/named_bak/
+    cp /var/named/chroot/var/named/*zone /var/named/chroot/var/named_bak/
+    echo "Remove zone config files"
+    rm -rf /var/named/chroot/var/named
+  else
+    echo "Backing up aaPanelDns config data..."
+    cp /var/named/chroot/etc/named.rfc1912.zones /var/named/chroot/etc/named.rfc1912.zones_aabak
+    echo "Backup config: named.rfc1912.zones"
+    mkdir -p /var/named/chroot/var/named_bak/
+    echo "Backup zone config files"
+    cp /var/named/chroot/var/named/*_aadef /var/named/chroot/var/named_bak/
+    cp /var/named/chroot/var/named/*zone /var/named/chroot/var/named_bak/
+  fi
+
   rm -f /www/server/panel/class_v2/ssl_dnsV2/aadns.pl
+  rm -f /www/server/panel/class_v2/ssl_dnsV2/aaDns_conf.json
+
 	/usr/bin/systemctl stop named-chroot
 	systemctl disable named-chroot
 	systemctl stop pdns
 	systemctl disable pdns
+	sleep 3
+	echo "aaPanelDns Service Success!"
 }
 
 if [ "${1}" == 'install' ];then
@@ -152,5 +166,5 @@ if [ "${1}" == 'install' ];then
 elif  [ "${1}" == 'update' ];then
 	update_DnsManager
 elif [ "${1}" == 'uninstall' ];then
-	Uninstall_DnsManager
+	Uninstall_DnsManager ${2}
 fi
